pyproject.toml
[tool.poetry]
name = "otodom-monitoring"
version = "0.1.0"
description = ""
authors = ["Sergey Ivanychev <sergeyivanychev@gmail.com>"]
readme = "README.md"
packages = [{include = "otodom_monitoring"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
furl = "^2.1.3"
typing-extensions = "^4.3.0"
pydantic = "^1.10.1"
click = "^8.1.3"
loguru = "^0.6.0"
toolz = "^0.12.0"
python-telegram-bot = "^13.13"


[tool.poetry.group.dev.dependencies]
ipython = "^8.4.0"
black = "^22.6.0"
isort = "^5.10.1"
pycln = "^2.1.1"
pyupgrade = "^2.37.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
